+++ config/routes.js
@@ -1,1 +1,2 @@
 Router.post('Menu#getMenu', '/api/menu/getMenu');
+Router.get('Menu#getRouteDocument', '/api/menu/route_document');